Re-use existing inset container
