+++ osu.Game.Rulesets.Mania/Skinning/LegacyBodyPiece.cs
@@ -6,10 +6,8 @@ using osu.Framework.Allocation;
 using osu.Framework.Bindables;
 using osu.Framework.Graphics;
 using osu.Framework.Graphics.Animations;
-using osu.Framework.Graphics.Containers;
 using osu.Framework.Graphics.OpenGL.Textures;
 using osu.Game.Rulesets.Mania.Objects.Drawables;
-using osu.Game.Rulesets.Mania.UI;
 using osu.Game.Rulesets.Objects.Drawables;
 using osu.Game.Rulesets.UI.Scrolling;
 using osu.Game.Skinning;
@@ -143,39 +141,6 @@ namespace osu.Game.Rulesets.Mania.Skinning
             base.Dispose(isDisposing);
 
             lightContainer?.Expire();
-        }
-
-        private class HitTargetInsetContainer : Container
-        {
-            private readonly IBindable<ScrollingDirection> direction = new Bindable<ScrollingDirection>();
-
-            protected override Container<Drawable> Content => content;
-            private readonly Container content;
-
-            private float hitPosition;
-
-            public HitTargetInsetContainer()
-            {
-                RelativeSizeAxes = Axes.Both;
-
-                InternalChild = content = new Container { RelativeSizeAxes = Axes.Both };
-            }
-
-            [BackgroundDependencyLoader]
-            private void load(ISkinSource skin, IScrollingInfo scrollingInfo)
-            {
-                hitPosition = skin.GetManiaSkinConfig<float>(LegacyManiaSkinConfigurationLookups.HitPosition)?.Value ?? Stage.HIT_TARGET_POSITION;
-
-                direction.BindTo(scrollingInfo.Direction);
-                direction.BindValueChanged(onDirectionChanged, true);
-            }
-
-            private void onDirectionChanged(ValueChangedEvent<ScrollingDirection> direction)
-            {
-                content.Padding = direction.NewValue == ScrollingDirection.Up
-                    ? new MarginPadding { Top = hitPosition }
-                    : new MarginPadding { Bottom = hitPosition };
-            }
         }
     }
 }
