+++ osu.Game.Rulesets.Mania/Skinning/LegacyStageBackground.cs
@@ -2,14 +2,12 @@
 // See the LICENCE file in the repository root for full licence text.
 
 using osu.Framework.Allocation;
-using osu.Framework.Bindables;
 using osu.Framework.Graphics;
 using osu.Framework.Graphics.Containers;
 using osu.Framework.Graphics.Shapes;
 using osu.Framework.Graphics.Sprites;
 using osu.Game.Rulesets.Mania.Beatmaps;
 using osu.Game.Rulesets.Mania.UI;
-using osu.Game.Rulesets.UI.Scrolling;
 using osu.Game.Skinning;
 using osuTK;
 using osuTK.Graphics;
@@ -143,39 +141,6 @@ namespace osu.Game.Rulesets.Mania.Skinning
                         }
                     }
                 };
-            }
-        }
-
-        private class HitTargetInsetContainer : Container
-        {
-            private readonly IBindable<ScrollingDirection> direction = new Bindable<ScrollingDirection>();
-
-            protected override Container<Drawable> Content => content;
-            private readonly Container content;
-
-            private float hitPosition;
-
-            public HitTargetInsetContainer()
-            {
-                RelativeSizeAxes = Axes.Both;
-
-                InternalChild = content = new Container { RelativeSizeAxes = Axes.Both };
-            }
-
-            [BackgroundDependencyLoader]
-            private void load(ISkinSource skin, IScrollingInfo scrollingInfo)
-            {
-                hitPosition = skin.GetManiaSkinConfig<float>(LegacyManiaSkinConfigurationLookups.HitPosition)?.Value ?? Stage.HIT_TARGET_POSITION;
-
-                direction.BindTo(scrollingInfo.Direction);
-                direction.BindValueChanged(onDirectionChanged, true);
-            }
-
-            private void onDirectionChanged(ValueChangedEvent<ScrollingDirection> direction)
-            {
-                content.Padding = direction.NewValue == ScrollingDirection.Up
-                    ? new MarginPadding { Top = hitPosition }
-                    : new MarginPadding { Bottom = hitPosition };
             }
         }
     }
